Services/AnomalyDetectionService/requirements.txt
rdflib
rdflib-jsonld
kafka-python
pandas
numpy
scikit-learn
pyprog
requests